requirements.txt
boto3==1.34.113
botocore==1.34.113
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
idna==3.7
jmespath==1.0.1
pyasn1==0.6.0
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.2
rsa==4.9
s3transfer==0.10.1
six==1.16.0
urllib3==2.2.1
